third_party/WebKit/LayoutTests/security/block-test-expected.txt

http://255.255.255.255:1/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:1/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:1/test.jpg
http://255.255.255.255:7/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:7/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:7/test.jpg
http://255.255.255.255:9/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:9/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:9/test.jpg
http://255.255.255.255:11/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:11/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:11/test.jpg
http://255.255.255.255:13/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:13/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:13/test.jpg
http://255.255.255.255:15/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:15/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:15/test.jpg
http://255.255.255.255:17/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:17/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:17/test.jpg
http://255.255.255.255:19/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:19/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:19/test.jpg
http://255.255.255.255:20/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:20/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:20/test.jpg
http://255.255.255.255:21/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:21/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:21/test.jpg
http://255.255.255.255:22/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:22/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:22/test.jpg
http://255.255.255.255:23/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:23/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:23/test.jpg
http://255.255.255.255:25/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:25/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:25/test.jpg
http://255.255.255.255:37/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:37/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:37/test.jpg
http://255.255.255.255:42/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:42/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:42/test.jpg
http://255.255.255.255:43/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:43/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:43/test.jpg
http://255.255.255.255:53/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:53/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:53/test.jpg
http://255.255.255.255:77/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:77/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:77/test.jpg
http://255.255.255.255:79/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:79/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:79/test.jpg
http://255.255.255.255:87/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:87/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:87/test.jpg
http://255.255.255.255:95/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:95/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:95/test.jpg
http://255.255.255.255:101/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:101/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:101/test.jpg
http://255.255.255.255:102/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:102/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:102/test.jpg
http://255.255.255.255:103/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:103/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:103/test.jpg
http://255.255.255.255:104/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:104/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:104/test.jpg
http://255.255.255.255:109/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:109/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:109/test.jpg
http://255.255.255.255:110/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:110/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:110/test.jpg
http://255.255.255.255:111/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:111/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:111/test.jpg
http://255.255.255.255:113/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:113/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:113/test.jpg
http://255.255.255.255:115/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:115/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:115/test.jpg
http://255.255.255.255:117/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:117/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:117/test.jpg
http://255.255.255.255:119/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:119/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:119/test.jpg
http://255.255.255.255:123/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:123/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:123/test.jpg
http://255.255.255.255:135/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:135/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:135/test.jpg
http://255.255.255.255:139/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:139/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:139/test.jpg
http://255.255.255.255:143/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:143/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:143/test.jpg
http://255.255.255.255:179/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:179/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:179/test.jpg
http://255.255.255.255:389/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:389/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:389/test.jpg
http://255.255.255.255:465/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:465/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:465/test.jpg
http://255.255.255.255:512/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:512/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:512/test.jpg
http://255.255.255.255:513/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:513/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:513/test.jpg
http://255.255.255.255:514/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:514/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:514/test.jpg
http://255.255.255.255:515/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:515/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:515/test.jpg
http://255.255.255.255:526/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:526/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:526/test.jpg
http://255.255.255.255:530/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:530/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:530/test.jpg
http://255.255.255.255:531/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:531/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:531/test.jpg
http://255.255.255.255:532/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:532/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:532/test.jpg
http://255.255.255.255:540/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:540/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:540/test.jpg
http://255.255.255.255:556/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:556/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:556/test.jpg
http://255.255.255.255:563/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:563/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:563/test.jpg
http://255.255.255.255:587/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:587/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:587/test.jpg
http://255.255.255.255:601/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:601/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:601/test.jpg
http://255.255.255.255:636/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:636/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:636/test.jpg
http://255.255.255.255:993/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:993/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:993/test.jpg
http://255.255.255.255:995/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:995/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:995/test.jpg
http://255.255.255.255:2049/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:2049/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:2049/test.jpg
http://255.255.255.255:3659/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:3659/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:3659/test.jpg
http://255.255.255.255:4045/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:4045/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:4045/test.jpg
http://255.255.255.255:6000/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6000/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6000/test.jpg
http://255.255.255.255:6665/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6665/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6665/test.jpg
http://255.255.255.255:6666/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6666/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6666/test.jpg
http://255.255.255.255:6667/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6667/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6667/test.jpg
http://255.255.255.255:6668/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6668/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6668/test.jpg
http://255.255.255.255:6669/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:6669/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:6669/test.jpg
http://255.255.255.255:65535/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:65535/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:65535/test.jpg
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:65536/test.jpg
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:4294967295/test.jpg
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:4294967296/test.jpg
CONSOLE WARNING: Subresource requests using legacy protocols (like `ftp:`) are blocked. Please deliver web-accessible resources over modern protocols like HTTPS. See https://www.chromestatus.com/feature/5709390967472128 for details.
CONSOLE MESSAGE: line 32: PASS: error event fired for ftp://255.255.255.255/test.jpg
CONSOLE MESSAGE: line 32: PASS: error event fired for ftp://255.255.255.255:21/test.jpg
CONSOLE MESSAGE: line 32: PASS: error event fired for ftp://255.255.255.255:22/test.jpg
http://255.255.255.255:0/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255:0/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:0/test.jpg
http://255.255.255.255/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:80/test.jpg
http://255.255.255.255/test.jpg - willSendRequest <NSURLRequest URL http://255.255.255.255/test.jpg, main document URL block-test.html, http method GET>
CONSOLE MESSAGE: line 32: PASS: error event fired for http://255.255.255.255:/test.jpg
This test attempts to change the src of an IMG tag to all black listed ports to confirm that each fires an error event. It also tries the FTP ports for exemptions.
